+++ NEWS.md
@@ -4,6 +4,7 @@ NEWS
 # vein 0.4.8 (Release date: 2021-04-07)
 
 - Adds extractor_raster
+- Adds wilcoxon test into verify (2021-04-25)
 
 # vein 0.4.7 (Release date: 2020-04-20)